refactored error logic
