+++ src/ContactsCard.ts
@@ -5,7 +5,7 @@ import ContactsModuleRdfLib from '@solid-data-modules/contacts-rdflib'
 import { saveNewGroup } from 'contacts-pane'
 import { clearPreviousMessage, complain, mention } from './buttonsHelper'
 import { contactExistsMessage, contactWasAddedSuccesMessage, errorContactCreation, errorGroupCreation, errorNotExistsAddressBookUri, groupIsRequired } from './texts'
-import { addErrorToErrorDisplay, checkAndAddErrorDisplay } from './contactsErrors'
+import { createErrorDisplaySection, addErrorToErrorDisplay, checkAndAddErrorToDisplay, checkAndRemoveErrorDisplay } from './contactsErrors'
 import { NamedNode } from 'rdflib'
 
 const CONTACTS_POPUP_OVERLAY_ID = 'contacts-popup-overlay'
@@ -293,48 +293,6 @@ const createGroupListSection = (
   } 
   groupListSection.appendChild(createAddressBookGroupCreationButton(context, contactsModule, addressBooksData, contactData))
   return groupListSection as HTMLElement
-}
-
-const createErrorDisplaySection = (
-  context: DataBrowserContext
-): HTMLElement => {
-  const setButtonOnClickHandler = (event) => {
-    event.preventDefault()
-    errorDisplaySection.classList.remove('contactsShowErrors')
-    const errorMessage = context.dom.getElementById('error-display-message')
-    if (errorMessage) errorMessage.textContent = ''
-  }
-
-  const errorDisplaySection = context.dom.createElement('section')
-  errorDisplaySection.setAttribute('role', 'alert')
-  errorDisplaySection.setAttribute('aria-label', 'Section to display error messages related to contact creation')
-  errorDisplaySection.setAttribute('id', 'error-display-section')
-  errorDisplaySection.classList.add('contactsErrorDisplay')
-
-  const closeButton = context.dom.createElement('button')
-  closeButton.setAttribute('type', 'button')
-  closeButton.classList.add('contactsCloseErrorDisplayButton')
-  closeButton.textContent = 'x'
-  closeButton.addEventListener('click', setButtonOnClickHandler)
-
-  const errorMessage = context.dom.createElement('p')
-  errorMessage.setAttribute('id', 'error-display-message')
-  errorMessage.classList.add('contactsErrorMessage')
-
-  errorDisplaySection.appendChild(closeButton)
-  errorDisplaySection.appendChild(errorMessage)
-  return errorDisplaySection
-}
-
-const checkAndRemoveErrorDisplay = (
-  context: DataBrowserContext
-) => {
-  const errorDisplaySection = context.dom.getElementById('error-display-section')
-  if (errorDisplaySection && errorDisplaySection.classList.contains('contactsShowErrors')) {
-    errorDisplaySection.classList.remove('contactsShowErrors')
-    const errorMessage = context.dom.getElementById('error-display-message')
-    if (errorMessage) errorMessage.textContent = ''
-  }
 }
 
 const createNewContactCreationButton = (
@@ -882,7 +840,7 @@ const createGroupButton = (
     
     if (previouslySelected) {
       selectedGroupButton.classList.remove('contactsSelectedButton', 'selectedGroup')
-      checkAndAddErrorDisplay(context, groupIsRequired)
+      checkAndAddErrorToDisplay(context, groupIsRequired)
     } else {
       selectedGroupButton.classList.add('contactsSelectedButton', 'selectedGroup') 
       checkAndRemoveErrorDisplay(context)
